[REFACTOR] Move asset preferences cleanup to common
